+++ src/apsg/plotting/_plot_artists.py
@@ -11,7 +11,7 @@ from apsg.feature._container import (
     Vector3Set,
 )
 from apsg.feature._geodata import Cone, Fault, Foliation, Pair
-from apsg.feature._tensor3 import Ellipsoid, Tensor3
+from apsg.feature._tensor3 import Ellipsoid, Stress3, Tensor3
 from apsg.math._vector import Vector3
 
 # StereoNet
@@ -284,6 +284,20 @@ class StereoNet_Tensor(StereoNet_Artists):
             self.kwargs["label"] = self.args[0].label()
 
 
+class StereoNet_Stress(StereoNet_Artists):
+    def __init__(self, factory, *args, **kwargs):
+        super().__init__(factory, *args, **kwargs)
+        self.stereonet_method = "_stress"
+        self.args = args[:1]  # take max 1 args
+        self.parse_kwargs(kwargs)
+
+    def parse_kwargs(self, kwargs):
+        super().update_kwargs("stereonet_default_stress_kwargs")
+        self.kwargs.update((k, kwargs[k]) for k in self.kwargs.keys() & kwargs.keys())
+        if not isinstance(self.kwargs["label"], str):
+            self.kwargs["label"] = self.args[0].label()
+
+
 class StereoNet_Contour(StereoNet_Artists):
     def __init__(self, factory, *args, **kwargs):
         super().__init__(factory, *args, **kwargs)
@@ -395,6 +409,13 @@ class StereoNetArtistFactory:
     def create_tensor(*args, **kwargs):
         if all([issubclass(type(arg), Tensor3) for arg in args[:1]]):
             return StereoNet_Tensor("create_tensor", *args, **kwargs)
+        else:
+            raise TypeError("Not valid arguments for Stereonet arrow")
+
+    @staticmethod
+    def create_stress(*args, **kwargs):
+        if all([issubclass(type(arg), Stress3) for arg in args[:1]]):
+            return StereoNet_Stress("create_stress", *args, **kwargs)
         else:
             raise TypeError("Not valid arguments for Stereonet arrow")